Move DashboardBlazeCardCellViewModel to a separate file
